doc: fail on errors when running asciidoctor
Asciidoctor defaults to continuing in spite of various warnings and
errors, for compatibility with Asciidoc.  In 1.5.7, an option was added
to control this behavior[1].  Use it to fail when errors are reported
during processing.

Asciidoctor 1.5.7 was released May 2, 2018.  That's ancient for some and
relatively new for others.  Debian oldstable has 1.5.4 (and 1.5.6.2 in
stretch-backports).  Similarly. RHEL-7 has 1.5.6.1.  It's probably
reasonable for such distributions to build with asciidoc instead or use
the pre-generated html/man docs.

[1] https://github.com/asciidoctor/asciidoctor/commit/53d9a2e5
    (resolves #2003 allow failure level to be specified using CLI
    option (PR #2674), 2018-04-20)

Signed-off-by: Todd Zullinger <tmz@pobox.com>