doc: add dependency on $(cmds_txt) to $(TXT_TO_*) targets
The `cmd-list.perl` script generates `cmds-$category.txt` files via the
cmd-list.made target.  The category files are included in git.txt (this
is currently the only user of those files).

If the perl script fails, the `cmds-$category.txt` files may be missing
or malformed.  Rather than catching the failure to complete the target,
we proceed to process additional targets with $(ASCIIDOC).

With Asciidoc, the missing include files cause a warning, yet the
command exits cleanly:

    asciidoc -f asciidoc.conf -amanversion=2.32.0 -amanmanual='Git Manual' \
             -amansource='Git' -b docbook -d manpage -o git.xml+ git.txt && \
    mv git.xml+ git.xml
    asciidoc: WARNING: git.txt: line 230: include file not found:
        /home/user/src/git/Documentation/cmds-mainporcelain.txt

... and then fails when running xmlto:

    xmlto -m manpage-normal.xsl -m manpage-bold-literal.xsl \
          -m manpage-base-url.xsl -m manpage-quote-apos.xsl man git.xml
    xmlto: /home/user/src/git/Documentation/git.xml does not validate (status 3)
    xmlto: Fix document syntax or use --skip-validation option

With Asciidoctor, the missing included files cause an error, but for
compatibility with Asciidoc, Asciidoctor exits cleanly.  However, with
Asciidoctor we need to pass the --skip-validation flag to xmlto, per
f6461b82b9 (Documentation: fix build with Asciidoctor 2, 2019-09-15).
So xmlto continues past the validation error which stops the build with
asciidoc:

    xmlto -m manpage-normal.xsl -m manpage-bold-literal.xsl \
          -m manpage-base-url.xsl -m manpage-quote-apos.xsl \
          --skip-validation -x manpage.xsl man git.xml

The resulting git.1 contains the errors logged by asciidoctor at the
point of the include directives:

   Main porcelain commands
       Unresolved directive in git.txt - include::cmds-mainporcelain.txt[]

In asciidoctor >= 1.5.7, there is a --failure-level=[LEVEL] argument
which can be used to better handle this situation[1].  That's likely
worth enabling, if we can tolerate requiring Asciidoctor 1.5.7 as a
minimum.

However, the root of this failure is due to not having a dependency on
$(cmds_txt) in the targets which process txt files that may contain the
`cmds-$category.txt` includes.

Add a dependency on $(cmds_txt) to the targets which call $(TXT_TO_XML)
and $(TXT_TO_HTML).  We currently only need this for the targets which
generate git.1 and git.html, but having the dependency in place for
other targets allows new `include::cmds-$category.txt[]` entries in
other documentation without having to remember to update the targets
used.

Skip the $(OBSOLETE_HTML) target, as those files aren't expected to
change.

[1] https://github.com/asciidoctor/asciidoctor/commit/53d9a2e5
    (resolves #2003 allow failure level to be specified using CLI
    option (PR #2674), 2018-04-20)

Signed-off-by: Todd Zullinger <tmz@pobox.com>