t/lib-gpg: use --with-colons when parsing gpgsm output
The output of `gpgsm -K` changed in gnupg-2.3¹, breaking the parsing
used by the GPGSM prereq.

Use the `--with-colons` options for stable, machine-parseable output.
This allows the grep/cut/tr pipeline (and the subsequent echo which
appends ' S relax') to be replaced with a single call to awk to create
the ${GNUPGHOME}/trustlist.txt file.

¹ https://dev.gnupg.org/rGe7d70923901e is the change in 2.3, while
  https://dev.gnupg.org/rG9c57de75cf36 is the similar change in 2.2.

  The latter says: Here in 2.2 we keep the string "fingerprint:" and no
  not change it to "sha1 fpr" as we did in master (2.3). (sic)

Signed-off-by: Todd Zullinger <tmz@pobox.com>