test-lib (test_set_port): add 'random' argument
When running multiple builds concurrently, tests which run daemons
sometimes conflict with each other, leading to spurious failures:

    ++ /usr/sbin/httpd -d '/tmp/git-t.ck9I/trash directory.t9118-git-svn-funky-branch-names/httpd' \
       -f /builddir/build/BUILD/git-2.37.2/t/lib-httpd/apache.conf -DDAV -DSVN -c 'Listen 127.0.0.1:9118' \
       -k start
    (98)Address already in use: AH00072: make_sock: could not bind to address 127.0.0.1:9118
    no listening sockets available, shutting down
    AH00015: Unable to open logs
    ++ test 1 -ne 0

Generate a random port number if the env var provided to `test_set_port`
is set to 'random' instead of a number.  Add a simple test helper to
allow this to be done more portably than it could with only POSIX shell
commands.

Helped-by: Ondřej Pohořelský <opohorel@redhat.com>
Signed-off-by: Todd Zullinger <tmz@pobox.com>