t/lib-git-svn: try harder to find a port
As with the previous commits, try harder to find an open port to avoid
intermittent failures on busy/shared build systems.

By default, we make 3 attempts.  This may be overridden by setting
GIT_TEST_START_SVNSERVE_TRIES to a different value.

Run svnserve in daemon mode and use 'test_atexit' to stop it.  This is
cleaner than running in the foreground with --listen-once and having to
manage the PID ourselves.

Signed-off-by: Todd Zullinger <tmz@pobox.com>