t/lib-gpg: fix GPGSSH_VERIFYTIME prereq
The `! (ssh-keygen -Y verify ...` call in the GPGSSH_VERIFYTIME prereq
is incomplete.  The `-Y verify` option expects the data to be verified
to be passed on stdin.

We do this in the earlier GPGSSH prereq, but when GPGSSH_VERIFYTIME was
added in 50992f96c5 (ssh signing: verify ssh-keygen in test prereq,
2021-12-09) this was apparently missed.

As the test is checking for failure, it's possible this was missed
because the test failed when run by a test harness where stdin was not
connected to a terminal, resulting in the ssh-keygen call failing
regardless of the validity of the ssh signing key.

Provide the message on stdin to allow the prereq to work as expected.

Signed-off-by: Todd Zullinger <tmz@pobox.com>